// apps/demo/lib/content/types.ts
export type FeatureIcon =
	| 'Package'
	| 'Layers'
	| 'Shield'
	| 'Activity'
	| 'Link'
	| 'Filter'
	| 'Zap'
	| 'Code'
	| 'Server'
	| 'Terminal'

export interface Feature {
	id: string
	icon: FeatureIcon
	title: string
	description: string
	highlight?: string
	link?: {
		text: string
		href: string
	}
}